refactor(hyperf): refactor label helpers and update version in Helm chart
The label helpers in hyperf/templates/_helpers.tpl have been refactored to provide better distinction between init and server deployment. This makes the usage of labels clearer and easier to manage in the different contexts. Additionally, the version of the Helm chart has been updated from 1.0.0 to 1.0.1 in hyperf/Chart.yaml and index.yaml files.